requirements.txt
mistralai>=0.0.7
httpx>=0.24.1
pikepdf>=8.7.1
python-dotenv>=1.0.0
ipython>=8.15.0
jupyter>=1.0.0
matplotlib>=3.7.2
numpy>=1.24.3
langchain>=0.0.335
langchain-openai>=0.0.2
langchain-google-genai>=1.0.0
langchain-cohere>=0.1.0
langchain-tavily<=0.2.0
langgraph>=0.0.15
langsmith>=0.0.75
pydantic>=2.5.2
typing-extensions>=4.8.0
networkx>=3.1
openai>=1.0.0
google-generativeai>=0.3.0
cohere>=4.0.0
langchain-community>=0.0.10
langchain-text-splitters>=0.0.1
tiktoken>=0.5.0
